Changes made as of 7-1-17 at 7:57pm
I have made drastic changes to the CSS. Not much in the HTML, but there are a few in the footer. Currently, to make everything set to a max-width, the backgrounds of the nav, section, and footer do not extend to the viewport's edge. I am saving that until later. I have been playing with that for too long.